Bash init added, kd incorrectly passed from motor0 to motor3

+++ CanCarController/pid.hpp
@@ -15,8 +15,8 @@ public:
     float max;
     float imax;
     float tgtspd; // rpm转子，出端RPM = /36
-    float err[4][2];
-    float dBuf[4][2];
+    int16_t err[4][2];
+    int16_t dBuf[4][2];
 
     int pTmp;
     int iTmp[4];
@@ -26,7 +26,9 @@ public:
     void init_pid(float p, float i, float d, float max, float imax, float tgtspd); // 可以用聚类判断电机，或者直接导入canRx, tgtspd单位是RPM
     int pidUpdate(int16_t curspd, int i);
 
-    pid(/* args */) {}
+    pid(/* args */)
+    {
+    }
     ~pid() {}
 };
